ml_backend/pyproject.toml
[project]
name = "ml-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "celery-types==0.23.0",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-unfold>=0.69.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "django-filter>=25.2",
    "drf-spectacular>=0.28.0",
    "gunicorn==23.0.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "django-storages>=1.14.6",
    "boto3>=1.40.76",
    "django-redis>=6.0.0",
]
